+++ src/main/java/spring/boot/webflu/ms/cliente/app/impl/TipoCuentaServiceImpl.java
@@ -1,5 +1,5 @@
-package spring.boot.webflu.ms.cliente.app.impl;
-
-public class TipoCuentaServiceImpl {
-
-}
+//package spring.boot.webflu.ms.cliente.app.impl;
+//
+//public class TipoCuentaServiceImpl {
+//
+//}
