+++ src/main/java/spring/boot/webflu/ms/cliente/app/service/impl/ClienteServiceImpl.java
@@ -1,6 +1,7 @@
 package spring.boot.webflu.ms.cliente.app.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
