+++ src/main/resources/application.yml
@@ -1,0 +1,5 @@
+spring:
+   application:
+    name: ms-cliente
+   config:
+     import: optional:configserver:http://localhost:8888
