+++ src/main/resources/bootstrap.properties
@@ -1,2 +1,2 @@
-spring.application.name=ms-cliente
-spring.cloud.config.uri=http://localhost:8888
+#spring.application.name=ms-cliente
+#spring.cloud.config.uri=http://localhost:8888
